+++ lib/presentation/router/app_router.dart
@@ -5,6 +5,7 @@ import 'package:tollgate_app/presentation/features/tollgate/payment/payment_scre
 import 'package:tollgate_app/presentation/features/settings/settings_screen.dart';
 
 import '../features/wallet/send/send_screen.dart';
+import '../features/wallet/receive/receive_screen.dart';
 import '../features/wallet/reserve/reserve_screen.dart';
 import '../features/wifi/connection_details/connection_screen.dart';
 import '../features/home/home_screen.dart';
@@ -124,6 +125,15 @@ GoRouter router() => GoRouter(
                       path: 'send',
                       pageBuilder: (context, state) => CustomTransitionPage(
                         child: const SendScreen(),
+                        transitionsBuilder:
+                            (context, animation, secondaryAnimation, child) =>
+                                _buildSlideLeftTransition(animation, child),
+                      ),
+                    ),
+                    GoRoute(
+                      path: 'receive',
+                      pageBuilder: (context, state) => CustomTransitionPage(
+                        child: const ReceiveScreen(),
                         transitionsBuilder:
                             (context, animation, secondaryAnimation, child) =>
                                 _buildSlideLeftTransition(animation, child),