+++ lib/presentation/router/routes.dart
@@ -8,5 +8,6 @@ abstract final class Routes {
   static const payment = '/payment';
   static const mint = '/wallet/mint';
   static const send = '/wallet/send';
+  static const receive = '/wallet/receive';
   static const reserve = '/wallet/reserve';
 }